submodule: Add vim-visual-star-search plugin.
File Search

This plugin allows you to select some text using Vim's visual mode, then
hit * and # to search for it elsewhere in the file. For example, hit V,
select a strange sequence of characters like "$! $!", and hit star.
You'll find all other runs of "$! $!" in the file.

By default, when you hit star while making a visual selection, vim
extends the selection to the next word that matches the word under the
cursor.

Recursive Search

If you hit <leader>* (\* unless you changed the mapleader), vim
recursively vimgreps for the word under the cursor or the visual
selection.

Signed-off-by: mengshengzhi <mengshengzhi@baidu.com>